Refactor database configurations and log options
The commit refactors the MySQL database configurations in the .env file,
updating the DB_PASS, DB_HOST, and DB_BASE values. Additionally, the
commit modifies the PhpSlides.php file, changing the $log property to be
of type bool and introduces a new static method config() for configuring
the request log. This change enhances the security and flexibility of
the application's database setup and logging functionality. No issues
are associated with this commit.

#Refactor #DatabaseConfigurations #LogOptions