Refactor route callback function in README, update copyright notice
- Change author name in LICENSE file from Joseph Onipede to Dave Conco
- Update code comments and examples in the README.md file to indicate
the use of closure functions instead of regular functions for GET routes

The commit updates the author's name in the LICENSE file to reflect a
change in ownership. Additionally, the README.md file is modified to
provide accurate information on the preferred method for defining GET
routes using closure functions instead of regular functions. This change
improves code clarity and provides accurate guidance for developers. No
issues are associated with this commit.

Note: This commit message assumes that the change in author name is
intentional and reflects an ownership update.